[UPDATED] Preferences page.
The checkbox form is dynamically generated with the map method applied
to an array of data imported from another file.
The data to be validated at submit are taken from the Local Storage and
no longer from useRef.
The checked state of the checkboxes and the related classes selected
refer to the Local Storage.