[UPDATED] Refactoring of Slider.js
The Component Slider no longer uses useRef and useEffect for assigning
selection and error classes. Classes are managed directly by jsx tags.
The reference values are no longer passed as props from the Component
Plans, but are taken from the Local Storage.
Tags related to slider values and ticks are rendered dynamically with
the .map() method.